refactor: Move error file to miden-lib, build only when possible (#970)
