+++ miden-lib/src/errors/mod.rs
@@ -1,0 +1,2 @@
+#[rustfmt::skip]
+pub mod tx_kernel_errors;
