+++ miden-tx/src/lib.rs
@@ -22,8 +22,8 @@ pub use verifier::TransactionVerifier;
 
 mod errors;
 pub use errors::{
-    tx_kernel_errors::TX_KERNEL_ERRORS, AuthenticationError, DataStoreError,
-    TransactionExecutorError, TransactionProverError, TransactionVerifierError,
+    AuthenticationError, DataStoreError, TransactionExecutorError, TransactionProverError,
+    TransactionVerifierError,
 };
 
 pub mod auth;
